Update rediscommand.h
Changed to int64 to fit full range of redis accepted input.

+++ common/rediscommand.h
@@ -51,7 +51,7 @@ public:
     void formatHDEL(const std::string& key, const std::vector<std::string>& fields);
 
     /* Format EXPIRE key field command */
-    void formatEXPIRE(const std::string& key, const int& ttl);
+    void formatEXPIRE(const std::string& key, const int64_t& ttl);
 
     const char *c_str() const;
